tests: wait until the key is actually there before deleting it
The delete_reply_keypair function silently ignores when the key does
not exist. We have to wait until the key exists before deleting it
otherwise it will show up after the function runs.

I had the opportunity of fixing this race because it was happening all
the time when running make test on my laptop.