securedrop-admin: refactor sdconfig as a python script
This is a replacement for the site-specific file update which
validates user input when the user hits enter. It is based on the
prompt-toolkit because it is well maintained. If a site-specific file
exists, it is loaded and replaces the defaults. The default value is
displayed for the user to be modified. The vi / emacs key bindings are
available, as they are in bash.

Ansible is not fit for user interaction.